Disable remote forms by default (#141)
When using `form_with` in Rails, it defaults to ajax submitting them.
The problem here is that the behavior is inconsistent when JS is enabled
or disabled. Flash messages don't display when JS is enabled, and there
needs to be a `.js.haml` template to update the page via JS if there is
any other update to be done. This causes some duplication between the
formats, so it would be nice to come up with a transparent way to make
both work with minimal duplication.