Set up Trello integration
A bunch of stuff in here. The integration is rudimentary, but it's a
proof of concept. The Trello checks out to be moved to a background job
and we should probably be using Trello webhooks rather than querying
their API every request. There's a bunch more that probably ought to be
happening here, too, but gotta start somewhere.