refactor user find logic, et al (#497)
I want to add support for an API key, but was finding the finder logic
for users was getting increasingly complicated. I've extracted to helper
classes `User::SessionFind` and `User::FindBy` in order to wrap the
`NullUser` logic rather than overriding the default Rails behavior. This
surfaced some other issues, so I also ended up adding an `admin` bit to
the user record and deleting code from `accounts_controller` which
doesn't appear to be reachable. Normally I'd break this up into a series
of PRs, but nobody is reading this so I get to do it the lazy way.