spawn background jobs for each check refresh
This will make it so that checks can run in parallel, but also make it
so that if one check raises an error, it doesn't block all of the checks
that follow it.

I'm also switching `strict_loading` to only log violations rather than
raising an error. It can be hard to detect them up front. Later on we
might hook the log into an error tracking service for easier followup.