Unnest target model (#544)
**What**

Unnests `Check::Target` everywhere to just `Target`.

**Why**

It's going to be easier to reason about classes in the app with a
flatter structure. I want to add a new `RefreshRate` class for targets,
which would add yet another nesting level if we were to maintain the
current paradigm: `Check::Target::RefreshRate`. I doubt this application
will ever reach a state where it will have so many base models it
becomes unwieldy to have them all in the top level folder.