Unnest count_refresh model (#545)
**What**

Renames references from `CheckCount` to `Count`.

**Why**

I don't think there's a great need to have the extra qualifier on the
name here. Currently we only have counts for the `Check` model, and I'm
not aware of any reason we would have any other. So for now I think we
can do without the extra letters.