RuboCop: switch to compact module syntax (#556)
**Why**

I want to namespace things under the model name, but it's annoying to
have a whole bunch of `< ApplicationRecord` all over the place. In many
cases I have seen people use the plural for the module name instead,
e.g.: `Targets::Refresh`. However, this feels like kind of a hack to get
around the naming issue.

There used to be weird module resolution issues with Ruby, but it looks
like they were resolved in more recent versions. Instead of dynamically
declaring a module, Ruby tries to look up the constant for each thing in
the chain, and raises an error when it is not found. As such, I needed
to explicitly declare a couple of modules, which actually strikes me as
an improvement.