remove refresh meta tag
Refreshing seems to be causing Trello throttling for some reason. There
are a couple of fixes that need to be made. For one, we shouldn't be
hitting Trello for each check every time we refresh the page. For
another, we should be refreshing checks only when they change, probably
via Turbo. We should also use Trello webhooks to trigger updates when
something changes, rather than manually polling frequently.